requirements.txt
requests==2.31.0
schedule==1.2.0
twilio==8.10.0
flask==3.0.2
python-dotenv==1.0.0
urllib3==2.0.7
googletrans==4.0.2